feat: refact javascript
